requirements.txt
python-telegram-bot==20.7
paramiko==3.4.0
cryptography==42.0.0
python-dotenv==1.0.0
aiosqlite==0.20.0
anthropic>=0.40.0
aiohttp>=3.9.0
